Update with-firebase-hosting example
Fix code
