+++ examples/with-firebase-hosting/app/components/App.js
@@ -1,11 +1,9 @@
-import Link from "next/link"
+import Header from "./Header"
 
-export default ({ pathname }) =>
-  <header>
-    <Link href="/">
-      <a className={pathname === "/" && "is-active"}>Home</a>
-    </Link>{" "}
-    <Link href="/about">
-      <a className={pathname === "/about" && "is-active"}>About</a>
-    </Link>
-  </header>
+const App = ({ children }) =>
+  <main>
+    <Header />
+    {children}
+  </main>
+
+export default App
