+++ examples/with-firebase-hosting/app/components/Header.js
@@ -1,9 +1,11 @@
-import Header from "./Header"
+import Link from "next/link"
 
-const App = ({ children }) =>
-  <main>
-    <Header pathname={"test"} />
-    {children}
-  </main>
-
-export default App
+export default ({ pathname }) =>
+  <header>
+    <Link href="/">
+      <a className={pathname === "/" && "is-active"}>Home</a>
+    </Link>{" "}
+    <Link href="/about">
+      <a className={pathname === "/about" && "is-active"}>About</a>
+    </Link>
+  </header>
